+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wormy Dirt texture by raspmary.
 - Deepslate Gravel and Raked Deepslate Gravel (textures by raspmary)
 - Blackstone Gravel and Raked Blackstone Gravel (textures by raspmary)
+- Copper Doors (Copper and Tuff Backport) can now be opened as double doors (Quark)
 
 ### Fixed
 - Further fixes for Thrashers.